Download icon from mibap repository
In a future version of this build pipeline, the icns file will be
generated on-the-fly by converting an existing png file from the
Inkscape repository. Until this has been implemented - and to avoid
introducing a 1 MiB blob to Inkscape's Git master - the icns file
will be downloaded from the mibap repository instead.